+++ src/main/java/ongi/ongibe/domain/album/controller/AlbumController.java
@@ -4,14 +4,19 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import ongi.ongibe.common.BaseApiResponse;
 import ongi.ongibe.domain.album.dto.*;
 import ongi.ongibe.domain.album.service.AlbumService;
-import ongi.ongibe.swagger.*;
+import ongi.ongibe.swagger.album.BaseApiResponse_AlbumDetailResponse;
+import ongi.ongibe.swagger.album.BaseApiResponse_AlbumInviteResponse;
+import ongi.ongibe.swagger.album.BaseApiResponse_AlbumMemberResponse;
+import ongi.ongibe.swagger.album.BaseApiResponse_AlbumOwnerTransferResponse;
+import ongi.ongibe.swagger.album.BaseApiResponse_AlbumSummaryResponseList;
+import ongi.ongibe.swagger.album.BaseApiResponse_MonthlyAlbumResponse;
+import ongi.ongibe.swagger.album.BaseApiResponse_String;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
